backend/requirements.txt
fastapi[standard]
uvicorn[standard]
SQLAlchemy>=2.0
alembic
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
pydantic-settings
psycopg2-binary>=2.9.9
